Refactor transformers to share more logic.
